+++ components/common/Tooltip.js
@@ -2,7 +2,7 @@ import React from 'react'
 import InfoIcon from '../../static/circle-information'
 import ReactTooltip from 'react-tooltip'
 
-export const Tooltip = props => {
+const Tooltip = props => {
   return (
     <>
       <span data-tip={props.dataTip}>
@@ -14,9 +14,4 @@ export const Tooltip = props => {
   )
 }
 
-export const TooltipDescriptions = {
-  ROADS: 'kilometers of road added or modified',
-  RAILWAYS: 'kilometers of railway added or modified',
-  COASTLINES: 'kilometers of coastline added or modified',
-  WATERWAYS: 'kilometers of waterways added or modified'
-}
+export default Tooltip