+++ components/users/AllUsersTable.js
@@ -3,6 +3,8 @@ import Link from '../Link'
 import { formatDecimal, formatEditTimeDescription } from '../../lib/utils/format'
 import { LoadingState } from '../common/LoadingState'
 import { parse } from 'date-fns'
+import TableHeaders from '../common/TableHeaders'
+import { tableHeaderNames } from '../../lib/enums'
 
 const UsersTable = ({ apiStatus, users }) => {
   let content = <div />
@@ -12,11 +14,7 @@ const UsersTable = ({ apiStatus, users }) => {
         <table>
           <thead>
             <tr>
-              <th>Rank</th>
-              <th>Name</th>
-              <th>Country</th>
-              <th>Total Edits</th>
-              <th>Last Edit</th>
+              <TableHeaders tableName={tableHeaderNames.ALL_USERS} />
             </tr>
           </thead>
           <tbody>