+++ pages/admin/users.js
@@ -8,6 +8,8 @@ import { isAdmin } from '../../lib/utils/roles'
 import NotLoggedIn from '../../components/NotLoggedIn'
 import AdminHeader from '../../components/admin/AdminHeader'
 import { LoadingState } from '../../components/common/LoadingState'
+import TableHeaders from '../../components/common/TableHeaders'
+import { tableHeaderNames } from '../../lib/enums'
 
 export class AdminUsers extends Component {
   constructor () {
@@ -41,6 +43,9 @@ export class AdminUsers extends Component {
     const { admin } = this.props
     if (!admin || !admin.users) return
 
+    // Reorders headers to show UserId first
+    const headers = <TableHeaders tableName={tableHeaderNames.ADMIN_USER} />
+
     return (
       <div className='admin'>
         <h1>All Users</h1>
@@ -48,9 +53,7 @@ export class AdminUsers extends Component {
           <table className='admin-table'>
             <thead>
               <tr>
-                <th>ID</th>
-                <th>Name</th>
-                <th>Roles</th>
+                {headers}
               </tr>
             </thead>
             <tbody>